Turn on keepCw by default
In Misskey, CW's are not automatically included in replies to a CW'ed
post unless the keepCw registry flag is set. This flag defaults to false
and there is not an easy user toggle for it.

To make it easier for users moving to Misskey from Mastodon, this
change enables the registry flag by default.